added error for IOS reloading to stub, fixed IOS reloading for system menu, changed settings to make it smaller (THIS REV IS NOT COMPATABLE WITH OLDER REVS. but the installer cleans settings anyway) , made the HBC stub not load when its already loaded (rare cases) . also moved the stub loading to before settings get loaded. and to stop confusions a bit : it says 0.3c now eventho there is no public release of 0.3c

+++ priiloader/source/gecko.cpp
@@ -21,7 +21,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 
 #include "gecko.h"
-u32 GeckoFound = 0;
+u8 GeckoFound = 0;
 
 void CheckForGecko( void )
 {
